Get rid of some questionable patterns
