+++ src/neurospin_to_bids/__main__.py
@@ -25,53 +25,11 @@ import pkg_resources
 from . import acquisition_db
 from . import bids
 from . import exp_info
-from .utils import DataError, UserError
+from . import utils
+from .utils import DataError, UserError, yes_no
 
 
 logger = logging.getLogger(__name__)
-
-
-NONINTERACTIVE = False
-
-
-def yes_no(question: str, *,
-           default: str = None, noninteractive: bool = None) -> bool:
-    """A simple yes/no prompt
-
-    Args:
-        question (str): The question to be answered.
-        default (bool, optional): Default answer to `question`.
-                                  Defaults to None.
-
-    Raises:
-        ValueError: Raise `ValueError` when default answer is not
-                    `yes` or `no`.
-
-    Returns:
-        bool: Boolean answer to the yes/no question.
-    """
-    valid = {"yes": True, "y": True, "no": False, "n": False}
-    if NONINTERACTIVE:
-        if noninteractive is not None:
-            return noninteractive
-        else:
-            return valid[default]
-    if default is None:
-        prompt = " [y/n] "
-    elif default == "yes":
-        prompt = " [Y/n] "
-    elif default == "no":
-        prompt = " [y/N] "
-    else:
-        raise ValueError(f"invalid default answer: '{default}'")
-
-    while True:
-        choice = input(question + prompt).lower()
-        if choice == '' and default is not None:
-            return valid[default]
-        if choice in valid:
-            return valid[choice]
-        print("Please respond with 'yes' or 'no' " "(or 'y' or 'n').\n")
 
 
 def bids_copy_events(behav_path='exp_info/recorded_events',
@@ -608,7 +566,6 @@ def bids_acquisition_download(data_root_path='',
 
 
 def main(argv=sys.argv):
-    global NONINTERACTIVE
     prog = os.path.basename(argv[0])
     if sys.version_info < (3, 6):
         sys.stderr.write(f'ERROR: {prog} needs Python 3.6 or later\n')
@@ -668,8 +625,8 @@ def main(argv=sys.argv):
 
     logger.info('Started %s', ' '.join(shlex.quote(arg) for arg in argv))
 
-    NONINTERACTIVE = args.noninteractive
-    acquisition_db.ACQUISITION_ROOT_PATH = args.acquisition_dir
+    utils.set_noninteractive(args.noninteractive)
+    acquisition_db.set_root_path(args.acquisition_dir)
 
     try:
         if args.autolist:
